Updates structure and fixes lingering indicator
Wrappes the indicator logic in a JavaScript object and moves the
HTML and show/hide-code to the Structure tool. By moving the code
we can trigger in the load-event of the work-iframe and by doing so
prevent the lingering indicator. The indicator is no longer visible
in the other tools.