Improves logging view and log modification
== Log view
Fixes a problem where the selected file and line count was not selected
on submit. Moves the refresh button closer to the other controller
elements. Makes the log view larger by default.

== Log modification
Adds some documentation to the log4.xml file that explaines the use cases
for the available appenders. Also adds a new "template" appender that is
used by the new log modification logic. If the template appender is not
present a Console appender will be used instead.

Adds possibility to set the logger's level to Trace, Fatal, and Off.

When a logger's level is changed the modification is now remebered by the
system. This enables a couple of features. (1) the active modifications
are displayed on the Application state view, (2) easier to revert the
modification through a UI element in the application state view, and (3)
the system can properly restore the logger to its original state.