Adds a general settings field in app. settings
The general settings is a way for the CMS users to specify application
wide settings. The settings can be read through the TemplateController
or using the the new tag management:generalSetting.

The values in the new property-set is defined using two fields. The first
is the primary field and is used like any other Properties-textarea combo.
The second field however is only used when in working-mode and overrides
settings in the first field if they share keys. This enables the user to
disable or change features when in working-mode.