Updates CKEditor to 4.3.2, and fixes
Removes the fix for high contrast (CK 4.3.2 fixes the issue).

Removes an ajax called made in WYSIWYG inline editing that caused problems
with caret position in the text.

Adds a new concept where Infoglue can augment the behavior of the CKEditor
without making changes to the CKEditor's config. The content of the new
file WYSIWYGSettings.js is included at the end of the reponse from
WYSIWYGProperties.action.