Better message on validation error in EOS
The update-content-attribute action now send the Validation message as an
error text when a validation error occurs. The JS receiving the error
display the error message to the user.

If the constraint exception is a required field exception the offending
field name is also included in the alert message (this is a new feature).