Makes it possible to disable filter-based Gzip
When using Tomcat to deliver static assets images are always gziped. It
might be desirable to disable this feature since there are other ways to
gzip resources today (using the Tomcat connector for example). This change
makes that possible. The setting is only available through the
build.properties file since changes this at run-time seems a bit
excessive.