Adds an asset optimization step to the dump asset method
When an asset is to be dump to disk it can now be passed to a OS level
executable before the dump is finished. The executable can do anything but
is expected to perform optimization (e.g. ImageMagick) on the asset. Only
PNGs and JPGs can use the new function as this time.